etl_pipeline/requirements.txt
Faker==37.4.0
greenlet==3.2.3
mysql-connector-python==9.3.0
numpy==2.3.0
pandas==2.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
